# apps/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginDto:
    properties:
      email:
        type: string
      password:
        type: string
      token:
        type: string
    type: object
  auth.LoginResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/auth.Model'
    type: object
  auth.Model:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      twofa_last_token:
        type: string
      twofa_secret:
        type: string
      twofa_status:
        type: boolean
      updatedAt:
        type: string
    type: object
  auth.RefreshTokenDto:
    properties:
      refreshToken:
        type: string
    type: object
  auth.RegisterDto:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.TwoFADisableRequestDto:
    description: TwoFADisableRequestDto is used to request 2FA disable
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.TwoFASetupRequestDto:
    description: TwoFASetupRequestDto is used to request 2FA setup
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.TwoFASetupResponseDto:
    description: TwoFASetupResponseDto is used to respond with 2FA setup info
    properties:
      provisioningUri:
        type: string
      secret:
        type: string
    type: object
  auth.TwoFAVerifyRequestDto:
    description: TwoFAVerifyRequestDto is used to verify a TOTP code
    properties:
      code:
        type: string
      email:
        type: string
    type: object
  auth.TwoFAVerifyResponseDto:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  auth.UpdatePasswordDto:
    description: UpdatePasswordDto is used for updating user password
    properties:
      currentPassword:
        type: string
      newPassword:
        type: string
    type: object
...
  utils.APIError:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  utils.ApiResponse-any:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  utils.ApiResponse-auth_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/auth.LoginResponse'
      message:
        type: string
    type: object
...
host: localhost:8033
info:
  contact: {}
  title: Peekaping API
  version: "1.0"
paths:
  /auth/2fa/disable:
    post:
      consumes:
      - application/json
      parameters:
      - description: 2FA disable request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.TwoFADisableRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Disable 2FA (TOTP) for user
      tags:
      - Auth
  /auth/2fa/setup:
    post:
      consumes:
      - application/json
      parameters:
      - description: 2FA setup request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.TwoFASetupRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TwoFASetupResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Enable 2FA (TOTP) for user
      tags:
      - Auth
  /auth/2fa/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: 2FA verify request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.TwoFAVerifyRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TwoFAVerifyResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.TwoFAVerifyResponseDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Verify 2FA (TOTP) code for user
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Login admin
      tags:
      - Auth
  /auth/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Password update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.UpdatePasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Update user password
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Register new admin
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"